added main component
