+++ src/index.js
@@ -1,6 +1,11 @@
+import { StrictMode } from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
 import "./styles.scss";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(<App />);
+root.render(
+    <StrictMode>
+        <App />
+    </StrictMode>
+);
